% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Arigon}
\alias{Arigon}
\title{Arigon dataset}
\format{A data frame with 80000 rows and 3 variables}
\usage{
data(Arigon)
}
\description{
A mock dataset containing meaningless events in a fictional state of Arigon (data overlays Oregon).
\itemize{
  \item latitude. Latitude of event.
  \item longitude. Longitude of event.
  \item date. Date of event.
}
}
\author{
LTC Steve Henderson and The Department of Systems Engineering at West Point
}
\keyword{datasets}

