% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{clean_crime}
\alias{clean_crime}
\title{Houston crime dataset}
\description{
Lightly cleaned Houston crime; no NA events included and all dates recognized by \code{pointdensity}; data from January 2010 to August 2010 geocoded with Google Maps and courtesy of \pkg{ggmap}
}
\author{
Houston Police Department, City of Houston
}
\references{
http://www.houstontx.gov/police/cs/stats2.htm
}
\keyword{datasets}

